% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_difference.R
\name{daily_difference}
\alias{daily_difference}
\title{daily_difference}
\usage{
daily_difference(df, symbol)
}
\arguments{
\item{df}{this should be a data frame}

\item{symbol}{this should be a character string with quotations}
}
\value{
this will return a data frame with the daily difference from the average growth for a specified ticker symbol. It will also print out the mean daily growth and the mean difference from daily growth at the end.
}
\description{
daily_difference
}
\examples{
daily_difference (stocks, "63MOON")
}
