% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errormetrics.R
\name{errormetrics}
\alias{errormetrics}
\title{errormetrics}
\usage{
errormetrics(reg.mod)
}
\arguments{
\item{reg.mod}{a regression model}
}
\value{
will return the MSE and MAD in a matrix for the regression model
}
\description{
errormetrics
}
\examples{
errormetrics(reg.mod)
}
