% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_growth.R
\name{highest_growth}
\alias{highest_growth}
\title{highest_growth}
\usage{
highest_growth(df, timestamp, n = 10)
}
\arguments{
\item{df}{this should be a data frame}

\item{timestamp}{this should be a time stamp in quotations}

\item{n}{this is the number of observations you want to return (ex. top 5, top 10, etc)}
}
\value{
this will return the top n growth from a specific date. Growth is calculated as the difference between that day's close and the previous close.
}
\description{
highest_growth
}
\examples{
highest_growth(stocks, "2017-12-29", n=5)
}
