% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPEbacktest.R
\name{MAPEbacktest}
\alias{MAPEbacktest}
\title{MAPEbacktest}
\usage{
MAPEbacktest(model, tsdataset, orig, h = 10, plot = FALSE)
}
\arguments{
\item{model}{this should be an arima model}

\item{tsdataset}{this should be a time series dataset}

\item{orig}{the origin point where you want to start backtesting}

\item{h}{length of the horizon to backtest}

\item{plot}{boolean to generate a plot or not}
}
\value{
this will run a recursive backtest for your dataset and return the MAPE values for each iteration. It can also generate a plot to display these values.
}
\description{
MAPEbacktest
}
\examples{
MAPEbacktest(arima.mod, data, orig=100, h=12, plot=TRUE)
}
