% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingBacktest.R
\name{movingBacktest}
\alias{movingBacktest}
\title{movingBacktest}
\usage{
movingBacktest(model, tsdataset, orig, h)
}
\arguments{
\item{model}{an arima model}

\item{tsdataset}{a time series dataset}

\item{orig}{the origin point of the backtest}

\item{h}{the length of the horizon}
}
\value{
this will return the error values for a moving window backtest
}
\description{
movingBacktest
}
\examples{
movingBacktest (arima.mod, data, orig=100, h=12)
}
