% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_obs.R
\name{num_obs}
\alias{num_obs}
\title{Number of Observations Function}
\usage{
num_obs(df, symbol)
}
\arguments{
\item{df}{this should be a data.frame}

\item{symbol}{this should be a character string with quotations}
}
\value{
returns the number of observations within the data frame for a stock of a specified ticker symbol
}
\description{
Number of Observations Function
}
\examples{
num_obs(stocks, "63MOONS")
}
