% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starbucks.R
\docType{data}
\name{starbucks}
\alias{starbucks}
\title{Nutrition information for Starbucks menu items, including food and drinks}
\format{
177 x 7 data frame
\describe{
\item{Calories}{Calories in the drink}
\item{Fat}{Fat in the drink}
\item{Carb}{Carbohydrates in the drink}
\item{Fiber}{Riber in the drink}
\item{Protein}{Protein in the drink}
\item{Sodium}{Sodium in the drink}
}
}
\usage{
starbucks
}
\description{
Nutrition information for Starbucks menu items, including food and drinks
}
\keyword{datasets}
